ci: run only on ubuntu and Julia 1.11
It seems entirely overkill to run on 3 different operating systems and
for two different Julia versions for a package that is still in heavy
development. Currently we just need the CI to be rapid and automate code
checking. There is nothing in this package that is specific to a given
OS, and we can worry about compat with Julia versions once we approach a
first stable version.