feat: modified EventList implementation
Simplified the logic for reading in data and let the libraries throw
their errors back to the user to handle. Implemented a very basic
`filter!` function as demonstration for how this structure can be used,
and to add a simple API for when GTIs are tied together with the event
list.